feat：将LLMDroid与mobsf相结合

+++ LLMDroid-Droidbot/droidbot/droidbot.py
@@ -52,7 +52,8 @@ class DroidBot(object):
                  humanoid=None,
                  ignore_ad=False,
                  replay_output=None,
-                 code_coverage: Literal['time', 'androlog', 'jacoco'] = 'androlog'
+                 code_coverage: Literal['time', 'androlog', 'jacoco'] = 'androlog',
+                 external_driver=False
                  ):
         """
         initiate droidbot with configurations
@@ -79,7 +80,8 @@ class DroidBot(object):
         self.timeout = timeout
         self.timer = None
         self.keep_env = keep_env
-        self.keep_app = keep_app
+        self.external_driver = external_driver
+        self.keep_app = True if external_driver else keep_app
 
         self.device: Device = None
         self.app = None
@@ -120,7 +122,8 @@ class DroidBot(object):
                 profiling_method=profiling_method,
                 master=master,
                 replay_output=replay_output,
-                code_coverage=code_coverage
+                code_coverage=code_coverage,
+                external_driver=external_driver
             )
         except Exception:
             import traceback
@@ -156,7 +159,10 @@ class DroidBot(object):
 
             if not self.enabled:
                 return
-            self.device.install_app(self.app)
+            if self.external_driver:
+                self.logger.info("External driver mode: skip app installation.")
+            else:
+                self.device.install_app(self.app)
 
             if not self.enabled:
                 return
@@ -196,11 +202,12 @@ class DroidBot(object):
             self.droidbox.stop()
         if self.device:
             self.device.disconnect()
-        if not self.keep_env:
+        if self.device and not self.keep_env:
             self.device.tear_down()
-        if not self.keep_app:
+        if self.device and not self.keep_app:
             self.device.uninstall_app(self.app)
-        if hasattr(self.input_manager.policy, "master") and \
+        if self.input_manager and self.input_manager.policy and \
+           hasattr(self.input_manager.policy, "master") and \
            self.input_manager.policy.master:
             import xmlrpc.client
             proxy = xmlrpc.client.ServerProxy(self.input_manager.policy.master)
